asdasd

+++ ClimateObservations/ClimateObservations/Models/Geolocation.cs
@@ -7,8 +7,8 @@ namespace ClimateObservations.Models
     public class Geolocation
     {
         public int Id { get; set; }
-        public float? Longitude { get; set; }
-        public float? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
         public int Area_id { get; set; }
         public List<Geolocation> Locations { get; set; } = new List<Geolocation>();
     }
